linera_rpc::config
Type Alias ValidatorInternalNetworkConfig

pub type ValidatorInternalNetworkConfig = ValidatorInternalNetworkPreConfig<NetworkProtocol>;

The network configuration for all shards.

Aliased Type

struct ValidatorInternalNetworkConfig {
    pub public_key: Secp256k1PublicKey,
    pub protocol: NetworkProtocol,
    pub shards: Vec<ShardConfig>,
    pub host: String,
    pub port: u16,
    pub block_exporters: Vec<ExporterServiceConfig>,
    pub metrics_port: u16,
}

Fields
public_key: Secp256k1PublicKey

The public key of the validator.

protocol: NetworkProtocol

The network protocol to use for all shards.

shards: Vec<ShardConfig>

The available shards. Each chain UID is mapped to a unique shard in the vector in a static way.

host: String

The host name of the proxy on the internal network (IP or hostname).

port: u16

The port the proxy listens on the internal network.

block_exporters: Vec<ExporterServiceConfig>

The server configurations for the linera-exporter. They can be used as optional locations to forward notifications to destinations other than the proxy, by the workers.

metrics_port: u16

The port of the proxy’s metrics endpoint.

Implementations

impl ValidatorInternalNetworkConfig

pub fn proxy_address(&self) -> String

pub fn exporter_addresses(&self) -> Vec<String>