Enum linera_rpc::RpcMessage

pub enum RpcMessage {

    BlockProposal(Box<BlockProposal>),
    Certificate(Box<HandleCertificateRequest>),
    LiteCertificate(Box<HandleLiteCertRequest<'static>>),
    ChainInfoQuery(Box<ChainInfoQuery>),
    DownloadBlobContent(Box<BlobId>),
    DownloadCertificateValue(Box<CryptoHash>),
    DownloadCertificate(Box<CryptoHash>),
    DownloadCertificates(Box<Vec<CryptoHash>>),
    BlobLastUsedBy(Box<BlobId>),
    VersionInfoQuery,
    GenesisConfigHashQuery,
    Vote(Box<LiteVote>),
    ChainInfoResponse(Box<ChainInfoResponse>),
    Error(Box<NodeError>),
    VersionInfoResponse(Box<VersionInfo>),
    GenesisConfigHashResponse(Box<CryptoHash>),
    DownloadBlobContentResponse(Box<BlobContent>),
    DownloadCertificateValueResponse(Box<CertificateValue>),
    DownloadCertificateResponse(Box<Certificate>),
    DownloadCertificatesResponse(Box<Vec<Certificate>>),
    BlobLastUsedByResponse(Box<CryptoHash>),
    CrossChainRequest(Box<CrossChainRequest>),

}

Variants

BlockProposal(Box<BlockProposal>)

Certificate(Box<HandleCertificateRequest>)

LiteCertificate(Box<HandleLiteCertRequest<'static>>)

ChainInfoQuery(Box<ChainInfoQuery>)

DownloadBlobContent(Box<BlobId>)

DownloadCertificateValue(Box<CryptoHash>)

DownloadCertificate(Box<CryptoHash>)

DownloadCertificates(Box<Vec<CryptoHash>>)

BlobLastUsedBy(Box<BlobId>)

VersionInfoQuery

GenesisConfigHashQuery

Vote(Box<LiteVote>)

ChainInfoResponse(Box<ChainInfoResponse>)

Error(Box<NodeError>)

VersionInfoResponse(Box<VersionInfo>)

GenesisConfigHashResponse(Box<CryptoHash>)

DownloadBlobContentResponse(Box<BlobContent>)

DownloadCertificateValueResponse(Box<CertificateValue>)

DownloadCertificateResponse(Box<Certificate>)

DownloadCertificatesResponse(Box<Vec<Certificate>>)

BlobLastUsedByResponse(Box<CryptoHash>)

CrossChainRequest(Box<CrossChainRequest>)

Implementations

impl RpcMessage

pub fn target_chain_id(&self) -> Option<ChainId>

Obtains the ChainId of the chain targeted by this message, if there is one.

Only inbound messages have target chains.

pub fn is_local_message(&self) -> bool

Wether this messsage is “local” i.e. will be executed locally on the proxy or if it’ll be proxied to the server.