linera_rpc::grpc
Module api

Modules

chain_info_result

Nested message and enum types in ChainInfoResult.

cross_chain_request

Nested message and enum types in CrossChainRequest.

notifier_service_client

Generated client implementations.

notifier_service_server

Generated server implementations.

pending_blob_result

Nested message and enum types in PendingBlobResult.

validator_node_client

Generated client implementations.

validator_node_server

Generated server implementations.

validator_worker_client

Generated client implementations.

validator_worker_server

Generated server implementations.

Structs

AccountOwner

AccountPublicKey

/ Public key of a Linera account. / Corresponding private key is used to sign block proposals.

AccountSignature

/ Signature created by a Linera account.

BlobContent

A blob of binary data.

BlobId

A content-addressed blob ID i.e. the hash of the BlobContent.

BlobIds

Content-addressed blob IDs i.e. the hashes of the BlobContents.

BlockHeight

BlockProposal

An authenticated proposal for a new block.

Certificate

A certified statement from the committee.

CertificatesBatchRequest

A request for a batch of certificates.

CertificatesBatchResponse

A batch of certificates.

ChainId

ChainInfoQuery

Request information on a chain.

ChainInfoResponse

Response to ChainInfoQuery

ChainInfoResult

A wrapper around ChainInfoResponse which contains a serialized error variant

ConfirmUpdatedRecipient

Acknowledge the height of the highest confirmed blocks communicated with UpdateRecipient.

CrateVersion

Information about the Linera crate version the validator is running

CrossChainRequest

An internal request between chains within a validator.

CryptoHash

HandleConfirmedCertificateRequest

A certified statement from the committee, together with other certificates required for execution.

HandlePendingBlobRequest

A request to handle a pending blob.

HandleTimeoutCertificateRequest

A certified statement from the committee, together with other certificates required for execution.

HandleValidatedCertificateRequest

A certified statement from the committee, together with other certificates required for execution.

LiteCertificate

A certified statement from the committee, without the value.

NetworkDescription

Notification

Notify that a chain has a new certified block or a new message.

PendingBlobRequest

A request for a pending blob.

PendingBlobResult

A requested pending blob, or an error.

SubscriptionRequest

A request for client to subscribe to notifications for a given ChainId

UpdateRecipient

Communicate a number of messages from the sender to the recipient. Messages must be given by increasing block height.

ValidatorPublicKey

/ Public key of a validator.

ValidatorSignature

/ Signature created by a validator.

VersionInfo

Information about the version of Linera the validator is running

Enums

CertificateKind