linera_rpc::grpc::api

Module validator_node_server

Generated server implementations.

Structs

ValidatorNodeServer

How to communicate with a validator or a local node.

Constants

SERVICE_NAME

Generated gRPC service name

Traits

ValidatorNode

Generated trait containing gRPC methods that should be implemented for use with ValidatorNodeServer.