linera_rpc::grpc::api::validator_node_server

Trait ValidatorNode

pub trait ValidatorNode:
    Send
    + Sync
    + 'static {
    type SubscribeStream: Stream<Item = Result<Notification, Status>> + Send + 'static;

    // Required methods
    fn handle_block_proposal<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlockProposal>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_lite_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<LiteCertificate>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_confirmed_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<HandleConfirmedCertificateRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_validated_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<HandleValidatedCertificateRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_timeout_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<HandleTimeoutCertificateRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_chain_info_query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ChainInfoQuery>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscriptionRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeStream>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_version_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<VersionInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_network_description<'life0, 'async_trait>(
        &'life0 self,
        request: Request<()>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<NetworkDescription>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_shard_info<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ChainId>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ShardInfo>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_blob<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlobId>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BlobContent>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_pending_blob<'life0, 'async_trait>(
        &'life0 self,
        request: Request<PendingBlobRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<PendingBlobResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn handle_pending_blob<'life0, 'async_trait>(
        &'life0 self,
        request: Request<HandlePendingBlobRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn upload_blob<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlobContent>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BlobId>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CryptoHash>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Certificate>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_certificates<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CertificatesBatchRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CertificatesBatchResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_certificates_by_heights<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DownloadCertificatesByHeightsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CertificatesBatchResponse>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn download_raw_certificates_by_heights<'life0, 'async_trait>(
        &'life0 self,
        request: Request<DownloadCertificatesByHeightsRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<RawCertificatesBatch>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blob_last_used_by<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlobId>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<CryptoHash>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn blob_last_used_by_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlobId>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<Certificate>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn missing_blob_ids<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlobIds>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BlobIds>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Generated trait containing gRPC methods that should be implemented for use with ValidatorNodeServer.

Required Associated Types

type SubscribeStream: Stream<Item = Result<Notification, Status>> + Send + 'static

Server streaming response type for the Subscribe method.

Required Methods

fn handle_block_proposal<'life0, 'async_trait>( &'life0 self, request: Request<BlockProposal>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Propose a new block.

fn handle_lite_certificate<'life0, 'async_trait>( &'life0 self, request: Request<LiteCertificate>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Process a certificate without value.

fn handle_confirmed_certificate<'life0, 'async_trait>( &'life0 self, request: Request<HandleConfirmedCertificateRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn handle_validated_certificate<'life0, 'async_trait>( &'life0 self, request: Request<HandleValidatedCertificateRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn handle_timeout_certificate<'life0, 'async_trait>( &'life0 self, request: Request<HandleTimeoutCertificateRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

fn handle_chain_info_query<'life0, 'async_trait>( &'life0 self, request: Request<ChainInfoQuery>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handle information queries for this chain.

fn subscribe<'life0, 'async_trait>( &'life0 self, request: Request<SubscriptionRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeStream>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Subscribe to notifications for a set of Chain Ids.

fn get_version_info<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<VersionInfo>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request the nodeโs version info.

fn get_network_description<'life0, 'async_trait>( &'life0 self, request: Request<()>, ) -> Pin<Box<dyn Future<Output = Result<Response<NetworkDescription>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request the network description seen by this node.

fn get_shard_info<'life0, 'async_trait>( &'life0 self, request: Request<ChainId>, ) -> Pin<Box<dyn Future<Output = Result<Response<ShardInfo>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Request shard information for a specific chain.

fn download_blob<'life0, 'async_trait>( &'life0 self, request: Request<BlobId>, ) -> Pin<Box<dyn Future<Output = Result<Response<BlobContent>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download a blob.

fn download_pending_blob<'life0, 'async_trait>( &'life0 self, request: Request<PendingBlobRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<PendingBlobResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download a blob that belongs to a pending block on the given chain.

fn handle_pending_blob<'life0, 'async_trait>( &'life0 self, request: Request<HandlePendingBlobRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Handle a blob that belongs to a pending block on the given chain.

fn upload_blob<'life0, 'async_trait>( &'life0 self, request: Request<BlobContent>, ) -> Pin<Box<dyn Future<Output = Result<Response<BlobId>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Upload a blob. Returns an error if the validator has not seen a certificate using this blob.

fn download_certificate<'life0, 'async_trait>( &'life0 self, request: Request<CryptoHash>, ) -> Pin<Box<dyn Future<Output = Result<Response<Certificate>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download a certificate.

fn download_certificates<'life0, 'async_trait>( &'life0 self, request: Request<CertificatesBatchRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CertificatesBatchResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download a batch of certificates.

fn download_certificates_by_heights<'life0, 'async_trait>( &'life0 self, request: Request<DownloadCertificatesByHeightsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<CertificatesBatchResponse>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

/ Download a batch of certificates by heights.

fn download_raw_certificates_by_heights<'life0, 'async_trait>( &'life0 self, request: Request<DownloadCertificatesByHeightsRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<RawCertificatesBatch>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Download a batch of certificates, in serialized form, by their heights.

fn blob_last_used_by<'life0, 'async_trait>( &'life0 self, request: Request<BlobId>, ) -> Pin<Box<dyn Future<Output = Result<Response<CryptoHash>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the hash of the Certificate that last used a blob.

fn blob_last_used_by_certificate<'life0, 'async_trait>( &'life0 self, request: Request<BlobId>, ) -> Pin<Box<dyn Future<Output = Result<Response<Certificate>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns the certificate that last used the blob.

fn missing_blob_ids<'life0, 'async_trait>( &'life0 self, request: Request<BlobIds>, ) -> Pin<Box<dyn Future<Output = Result<Response<BlobIds>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Return the BlobIds that are not contained as Blob.