linera_rpc::grpc::api::validator_relay_server

Trait ValidatorRelay

pub trait ValidatorRelay:
    Send
    + Sync
    + 'static {
    // Required methods
    fn relay_lite_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RelayLiteCertificateRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn relay_confirmed_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RelayConfirmedCertificateRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn relay_chain_info_query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RelayChainInfoQueryRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn relay_upload_blob<'life0, 'async_trait>(
        &'life0 self,
        request: Request<RelayUploadBlobRequest>,
    ) -> Pin<Box<dyn Future<Output = Result<Response<BlobId>, Status>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ValidatorRelayServer.

Required Methods

fn relay_lite_certificate<'life0, 'async_trait>( &'life0 self, request: Request<RelayLiteCertificateRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Forward a lite certificate to another validator.

fn relay_confirmed_certificate<'life0, 'async_trait>( &'life0 self, request: Request<RelayConfirmedCertificateRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Forward a confirmed certificate to another validator.

fn relay_chain_info_query<'life0, 'async_trait>( &'life0 self, request: Request<RelayChainInfoQueryRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Ask another validator for the state of a chain.

fn relay_upload_blob<'life0, 'async_trait>( &'life0 self, request: Request<RelayUploadBlobRequest>, ) -> Pin<Box<dyn Future<Output = Result<Response<BlobId>, Status>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Upload a blob to another validator.