Struct linera_rpc::grpc::api::validator_worker_server::ValidatorWorkerServer

pub struct ValidatorWorkerServer<T> { /* private fields */ }

Interface provided by each physical shard (aka โworkerโ) of a validator or a local node.

All commands return either the current chain info or an error.

Repeating commands produces no changes and returns no error.

Some handlers may return cross-chain requests, that is, messages to be communicated to other workers of the same validator.

Implementations

impl<T> ValidatorWorkerServer<T>

pub fn new(inner: T) -> Self

pub fn from_arc(inner: Arc<T>) -> Self

pub fn with_interceptor<F>( inner: T, interceptor: F, ) -> InterceptedService<Self, F>
where F: Interceptor,

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing requests with the given encoding.

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress responses with the given encoding, if the client supports it.

pub fn max_decoding_message_size(self, limit: usize) -> Self

Limits the maximum size of a decoded message.

Default: 4MB

pub fn max_encoding_message_size(self, limit: usize) -> Self

Limits the maximum size of an encoded message.

Default: usize::MAX