Trait linera_rpc::grpc_network::grpc::notifier_service_server::NotifierService

pub trait NotifierService: Send + Sync + 'static {
    fn notify<'life0, 'async_trait>(
        &'life0 self,
        request: Request<Notification>
    ) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with NotifierServiceServer.

Required Methods

fn notify<'life0, 'async_trait>(
&'life0 self,
request: Request<Notification>
) -> Pin<Box<dyn Future<Output = Result<Response<()>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Notify a client that there has been an update for a ChainId.