Struct linera_rpc::grpc_network::grpc::validator_node_server::ValidatorNodeServer

pub struct ValidatorNodeServer<T: ValidatorNode> { /* private fields */ }

How to communicate with a validator or a local node.

Implementations

impl<T: ValidatorNode> ValidatorNodeServer<T>

pub fn new(inner: T) -> Self

pub fn from_arc(inner: Arc<T>) -> Self

pub fn with_interceptor<F>(
inner: T,
interceptor: F
) -> InterceptedService<Self, F>where
F: Interceptor,

pub fn accept_compressed(self, encoding: CompressionEncoding) -> Self

Enable decompressing requests with the given encoding.

pub fn send_compressed(self, encoding: CompressionEncoding) -> Self

Compress responses with the given encoding, if the client supports it.