Trait linera_rpc::grpc_network::grpc::validator_node_server::ValidatorNode

pub trait ValidatorNode: Send + Sync + 'static {
    type SubscribeStream: Stream<Item = Result<Notification, Status>> + Send + 'static;

    fn handle_block_proposal<'life0, 'async_trait>(
        &'life0 self,
        request: Request<BlockProposal>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_lite_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<LiteCertificate>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_certificate<'life0, 'async_trait>(
        &'life0 self,
        request: Request<CertificateWithDependencies>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_chain_info_query<'life0, 'async_trait>(
        &'life0 self,
        request: Request<ChainInfoQuery>
    ) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn subscribe<'life0, 'async_trait>(
        &'life0 self,
        request: Request<SubscriptionRequest>
    ) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeStream>, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Generated trait containing gRPC methods that should be implemented for use with ValidatorNodeServer.

Required Associated Types

type SubscribeStream: Stream<Item = Result<Notification, Status>> + Send + 'static

Server streaming response type for the Subscribe method.

Required Methods

fn handle_block_proposal<'life0, 'async_trait>(
&'life0 self,
request: Request<BlockProposal>
) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Propose a new block.

fn handle_lite_certificate<'life0, 'async_trait>(
&'life0 self,
request: Request<LiteCertificate>
) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Process a certificate without value.

fn handle_certificate<'life0, 'async_trait>(
&'life0 self,
request: Request<CertificateWithDependencies>
) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Process a certificate.

fn handle_chain_info_query<'life0, 'async_trait>(
&'life0 self,
request: Request<ChainInfoQuery>
) -> Pin<Box<dyn Future<Output = Result<Response<ChainInfoResult>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Handle information queries for this chain.

fn subscribe<'life0, 'async_trait>(
&'life0 self,
request: Request<SubscriptionRequest>
) -> Pin<Box<dyn Future<Output = Result<Response<Self::SubscribeStream>, Status>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: 'async_trait,

Subscribe to notifications for a set of Chain Ids.