linera_rpc::propagation

Trait HasOtelContext

pub trait HasOtelContext {
    // Required method
    fn get_otel_context(&self) -> Option<&ExtractedOtelContext>;
}

Trait for request types that can provide access to extracted OpenTelemetry context.

This trait abstracts over http::Request and tonic::Request to allow generic functions that work with either request type.

Required Methods

fn get_otel_context(&self) -> Option<&ExtractedOtelContext>

Returns the extracted OpenTelemetry context, if present.