linera_rpc::config

Type Alias ValidatorInternalNetworkConfig

pub type ValidatorInternalNetworkConfig = ValidatorInternalNetworkPreConfig<NetworkProtocol>;

The network configuration for all shards.

Aliased Type

struct ValidatorInternalNetworkConfig {
    pub public_key: Secp256k1PublicKey,
    pub protocol: NetworkProtocol,
    pub shards: Vec<ShardConfig>,
    pub block_exporters: Vec<ExporterServiceConfig>,
    pub proxies: Vec<ProxyConfig>,
}

Fields
public_key: Secp256k1PublicKey

The public key of the validator.

protocol: NetworkProtocol

The network protocol to use internally.

shards: Vec<ShardConfig>

The available shards. Each chain UID is mapped to a unique shard in the vector in a static way.

block_exporters: Vec<ExporterServiceConfig>

The server configurations for the linera-exporter. They can be used as optional locations to forward notifications to destinations other than the proxy, by the workers.

proxies: Vec<ProxyConfig>

The available proxies.

Implementations

impl ValidatorInternalNetworkConfig

pub fn exporter_addresses(&self) -> Vec<String>