Type Alias linera_rpc::config::ValidatorInternalNetworkConfig

pub type ValidatorInternalNetworkConfig = ValidatorInternalNetworkPreConfig<NetworkProtocol>;

The network configuration for all shards.

Aliased Type

struct ValidatorInternalNetworkConfig {
    pub name: ValidatorName,
    pub protocol: NetworkProtocol,
    pub shards: Vec<ShardConfig>,
    pub host: String,
    pub port: u16,
    pub metrics_host: String,
    pub metrics_port: u16,
}

Fields
name: ValidatorName

The name of the validator.

protocol: NetworkProtocol

The network protocol to use for all shards.

shards: Vec<ShardConfig>

The available shards. Each chain UID is mapped to a unique shard in the vector in a static way.

host: String

The host name of the proxy on the internal network (IP or hostname).

port: u16

The port the proxy listens on on the internal network.

metrics_host: String

The host name of the proxy’s metrics endpoint.

metrics_port: u16

The port of the proxy’s metrics endpoint.

Implementations

impl ValidatorInternalNetworkConfig

pub fn proxy_address(&self) -> String