Enum linera_rpc::RpcMessage

pub enum RpcMessage {

    BlockProposal(Box<BlockProposal>),
    TimeoutCertificate(Box<HandleTimeoutCertificateRequest>),
    ValidatedCertificate(Box<HandleValidatedCertificateRequest>),
    ConfirmedCertificate(Box<HandleConfirmedCertificateRequest>),
    LiteCertificate(Box<HandleLiteCertRequest<'static>>),
    ChainInfoQuery(Box<ChainInfoQuery>),
    DownloadBlobContent(Box<BlobId>),
    DownloadConfirmedBlock(Box<CryptoHash>),
    DownloadCertificates(Vec<CryptoHash>),
    BlobLastUsedBy(Box<BlobId>),
    MissingBlobIds(Box<Vec<BlobId>>),
    VersionInfoQuery,
    GenesisConfigHashQuery,
    Vote(Box<LiteVote>),
    ChainInfoResponse(Box<ChainInfoResponse>),
    Error(Box<NodeError>),
    VersionInfoResponse(Box<VersionInfo>),
    GenesisConfigHashResponse(Box<CryptoHash>),
    DownloadBlobContentResponse(Box<BlobContent>),
    DownloadConfirmedBlockResponse(Box<ConfirmedBlock>),
    DownloadCertificatesResponse(Vec<ConfirmedBlockCertificate>),
    BlobLastUsedByResponse(Box<CryptoHash>),
    MissingBlobIdsResponse(Box<Vec<BlobId>>),
    CrossChainRequest(Box<CrossChainRequest>),

}

Variants

BlockProposal(Box<BlockProposal>)

TimeoutCertificate(Box<HandleTimeoutCertificateRequest>)

ValidatedCertificate(Box<HandleValidatedCertificateRequest>)

ConfirmedCertificate(Box<HandleConfirmedCertificateRequest>)

LiteCertificate(Box<HandleLiteCertRequest<'static>>)

ChainInfoQuery(Box<ChainInfoQuery>)

DownloadBlobContent(Box<BlobId>)

DownloadConfirmedBlock(Box<CryptoHash>)

DownloadCertificates(Vec<CryptoHash>)

BlobLastUsedBy(Box<BlobId>)

MissingBlobIds(Box<Vec<BlobId>>)

VersionInfoQuery

GenesisConfigHashQuery

Vote(Box<LiteVote>)

ChainInfoResponse(Box<ChainInfoResponse>)

Error(Box<NodeError>)

VersionInfoResponse(Box<VersionInfo>)

GenesisConfigHashResponse(Box<CryptoHash>)

DownloadBlobContentResponse(Box<BlobContent>)

DownloadConfirmedBlockResponse(Box<ConfirmedBlock>)

DownloadCertificatesResponse(Vec<ConfirmedBlockCertificate>)

BlobLastUsedByResponse(Box<CryptoHash>)

MissingBlobIdsResponse(Box<Vec<BlobId>>)

CrossChainRequest(Box<CrossChainRequest>)

Implementations

impl RpcMessage

pub fn target_chain_id(&self) -> Option<ChainId>

Obtains the ChainId of the chain targeted by this message, if there is one.

Only inbound messages have target chains.

pub fn is_local_message(&self) -> bool

Wether this messsage is “local” i.e. will be executed locally on the proxy or if it’ll be proxied to the server.