Module linera_rpc::grpc

Modules

api

pool

transport

Structs

GrpcClient

GrpcNodeProvider

GrpcPrometheusMetricsMiddlewareLayer

GrpcPrometheusMetricsMiddlewareService

GrpcServer

GrpcServerHandle

Enums

GrpcError

GrpcProtoConversionError

Constants

GRPC_CHUNKED_MESSAGE_FILL_LIMIT

Limit of gRPC message size up to which we will try to populate with data when estimating. We leave 30% of buffer for the rest of the message and potential underestimation.

GRPC_MAX_MESSAGE_SIZE

Traits

GrpcProxyable

Types which are proxyable and expose the appropriate methods to be handled by the GrpcProxy